Refactored functions to be pure - taking alphabet as argument
Pure functions help keep the code base simple and are much easier to reason
about as all the inputs to the function are given.

Adding the alphabet as an argument makes the functions pure and it also enables
the use of multiple alphabets in the design.

The refactored code seems to suggest the conversion could be expressed more
elegantly as both functions take alphabet as an argument.