Refactor converter into one function
Simplified the code by combining the two functions.  As character-clack only
retrieved the value of a key, then it was simple to just use a get function call
rather than defining a separate function.

The code is a lot shorter and in this instance is felt just as readable, perhaps
even more readable.